feat: useMessageOperations optimization
